Conflicts:
	mythtv/libs/libmythbase/mythversion.h

+++ mythtv/libs/libmythui/mythuibuttontree.cpp
@@ -113,7 +113,7 @@ void MythUIButtonTree::SetTreeState(bool refreshAll)
     m_visibleLists = 0;
     uint listid = 0;
 
-    while (listid < m_buttonlists.count())
+    while (listid < (uint)m_buttonlists.count())
     {
         MythUIButtonList *list = m_buttonlists.at(listid);
